refactor: move Hardware settings into Advanced tab
Drops the dedicated /settings/hardware route and sidebar entry and renders
the compute-device selector, GPU list and warnings as an extra section
inside Settings → Advanced. Keeps the feature available but gated as
power-user territory, matching the maintainer's direction on PR #139.

Also tightens the extracted section: toStored preserves existing
compute.threads, hasAnyGpuBackend filters to known GPU backends only
(cuda/metal/vulkan/rocm) instead of treating openblas as a GPU, and
Refresh now calls utils.hardware.getSnapshot.fetch({ refresh: true }) so
tapping it actually forces re-detection.
