+++ README.md
@@ -34,7 +34,7 @@
 - [🎗 License](#-license)
 
 See also [docs/hardware-selection.md](docs/hardware-selection.md) for the
-new Settings → Hardware tab (CPU / GPU selector) and
+new Hardware section under Settings → Advanced (CPU / GPU selector) and
 [docs/building-whisper-cuda-windows.md](docs/building-whisper-cuda-windows.md)
 for building the CUDA whisper binary on Windows.
