--- apps/desktop/src/renderer/main/routes/settings/hardware.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import HardwareSettingsPage from "../../pages/settings/hardware";
-
-export const Route = createFileRoute("/settings/hardware")({
-  component: HardwareSettingsPage,
-});
