+++ apps/desktop/src/renderer/main/routes/settings/route.tsx
@@ -59,7 +59,6 @@ function SettingsLayoutContent() {
       "/settings/vocabulary": "Vocabulary",
       "/settings/shortcuts": "Shortcuts",
       "/settings/ai-models": "AI Models",
-      "/settings/hardware": "Hardware",
       "/settings/history": "History",
       "/settings/advanced": "Advanced",
       "/settings/about": "About",
